+++ src/main/java/com/example/taskmanagementservice/task/request/UpdateTaskStatusRequest.java
@@ -10,7 +10,7 @@ import lombok.Value;
 public class UpdateTaskStatusRequest {
 
     @NotNull
-    Integer taskId;
+    Long taskId;
 
     @NotNull(message = "Status is mandatory")
     TaskStatus status;
